+++ CHANGELOG.md
@@ -6,12 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2024-02-11
+
 ### Added
+
 - Draft message for each chat.
-- Display of group descriptions.
-- Reception of typing indicators.
+- Display of group descriptions and contact about information.
+- Reception of typing indicators (contact-only for now).
 
 ### Changed
+
 - The sidebar now highlights the active chat.
 - Grab focus of the entry when replying to a message.
 - UI improvements to the channel information dialog, which now also displays phone number, disappearing messages timer and description.
@@ -500,7 +504,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed cargo clippy.
 - Updated presage to official repository.
 
-[Unreleased]: https://gitlab.com/schmiddi-on-mobile/flare/-/compare/0.11.2...master
+[Unreleased]: https://gitlab.com/schmiddi-on-mobile/flare/-/compare/0.12.0...master
+[0.12.0]: https://gitlab.com/schmiddi-on-mobile/flare/-/compare/0.11.2...0.12.0
 [0.11.2]: https://gitlab.com/schmiddi-on-mobile/flare/-/compare/0.11.1...0.11.2
 [0.11.1]: https://gitlab.com/schmiddi-on-mobile/flare/-/compare/0.11.0...0.11.1
 [0.11.0]: https://gitlab.com/schmiddi-on-mobile/flare/-/compare/0.10.0...0.11.0